requirements.txt
PyQt6>=6.6.0
opencv-python>=4.9.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
mss>=9.0.0
PyAutoGUI>=0.9.54
PyYAML>=6.0
numpy>=1.26.0
Pillow>=10.0.0
pytest>=8.0.0
